// mobile-app/src/modules/loans/project-finance/types/projectFinance.types.ts
export interface MarketDetailsForm {
  targetMarket: string;
  marketType: string;
  targetGeography: string;
  customerSegment: string;
  expectedMarketShare: string;
  majorCompetitors: string;
  competitiveAdvantage: string;
}